requirements.txt
Pillow==8.3.2
ExifRead==2.3.2